requirements.txt
alembic==1.0.5
astroid==2.1.0
bcrypt==3.1.5
cffi==1.11.5
Click==7.0
Flask==1.0.2
Flask-Bcrypt==0.7.1
Flask-Login==0.4.1
Flask-Migrate==2.3.1
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
gunicorn==19.9.0
isort==4.3.4
itsdangerous==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
Mako==1.0.7
MarkupSafe==1.1.0
mccabe==0.6.1
Pillow==5.3.0
pycparser==2.19
pylint==2.2.2
python-dateutil==2.7.5
python-editor==1.0.3
six==1.12.0
SQLAlchemy==1.2.15
typed-ast==1.1.0
Werkzeug==0.14.1
wrapt==1.10.11
WTForms==2.2.1
WTForms-SQLAlchemy==0.1
